requirements.txt
fastapi
uvicorn
numpy
opencv-python
tensorflow
pydantic
gdown
python-dotenv
